cassandra-app/requirements.txt
cassandra-driver==3.6.0
click==6.6
cql==1.4.0
cqlsh==5.0.3
Flask==0.11.1
futures==3.0.5
itsdangerous==0.24
Jinja2==2.8
linecache2==1.0.0
MarkupSafe==0.23
simplejson==3.8.2
six==1.10.0
thrift==0.9.3
traceback2==1.4.0
unittest2==1.1.0
Werkzeug==0.11.10
